test: don't modify an empty file
First actual test! Since the plugin does not have any functionallity currently,
this test will ensure that the plugin does not add content to an empty file.

Added a few functions to lessen the amount of code in individual tests.
These functions can be found in common.ts.

The code for the tests is heavily based on the code from the plugin-pug repo.